:star2: feat: separate backend init
