+++ package.json
@@ -14,6 +14,7 @@
     "core-js": "^2.6.5",
     "firebase": "^6.0.0",
     "markdown-it": "^8.4.2",
+    "toml": "^3.0.0",
     "vue": "^2.6.6",
     "vue-router": "^3.0.1",
     "vuetify": "^1.5.9",
@@ -22,7 +23,7 @@
   "devDependencies": {
     "@vue/cli-plugin-babel": "^3.5.0",
     "@vue/cli-plugin-eslint": "^3.5.0",
-    "@vue/cli-service": "^3.5.0",
+    "@vue/cli-service": "^3.6.0",
     "@vue/eslint-config-standard": "^4.0.0",
     "babel-eslint": "^10.0.1",
     "eslint": "^5.8.0",
