+++ src/config.js
@@ -4,10 +4,17 @@ const config = { // needs to be ES6 module so it can be imported by webpack
   blips: {
     titleCutOff: 20 // cut off title after N characters (display only)
   },
-  firebase: {
+  backend: {
+    type: 'firebase',
     key: 'AIzaSyC3FxfpYywy5ZYpiWuf9nw8_vlbxibpQH8',
     project: 'techradar-f5834'
   },
+  /*
+  backend: {
+    type: 'file',
+    path: '../blips.toml'
+  },
+  */
   categories: ['Tools', 'Techniques', 'Platforms', 'Frameworks'], // quadrant 1-4
   states: ['Novice', 'Intermediate', 'Advanced', 'Veteran'], // should be 0 - 3 for tech radar, 4 for in use, 5 for no longer in use
   // personal proficiency level https://hr.nih.gov/working-nih/competencies/competencies-proficiency-scale
